Moved from spikesort to cluster subpackage.
