+++ phy/cluster/manual/tests/test_clustering.py
@@ -12,7 +12,7 @@ import numpy as np
 from numpy.testing import assert_array_equal
 from pytest import raises
 
-from ...datasets.mock import artificial_spike_clusters
+from ....datasets.mock import artificial_spike_clusters
 from ..clustering import Clustering, History
